azure-fraud-detector/cloudservice/requirements.txt
streamlit
azure-storage-blob
azure-search-documents==11.4.0
python-dotenv
